appcodes/requirements.txt
azure-identity==1.25.1
azure-ai-projects==2.0.0b3
azure-monitor-opentelemetry==1.8.2
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
openai==2.15.0
httpx==0.28.1
msal==1.34.0
azure.search.documents==11.6.0
